--- NewTank/Assets/Tanks/Scripts.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: fd9c214f640fae348958e74eaea6f543
-folderAsset: yes
-timeCreated: 1431009955
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
