--- NewTank/Assets/Tanks/Scripts/Camera.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 7a43051ea413c884fb05edbbb60d7cc3
-folderAsset: yes
-timeCreated: 1431010000
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
