--- NewTank/Assets/Tanks/Scripts/Camera/CameraControl.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: c54e3c1f18f8cbb49b10f5e2dc8a1865
-timeCreated: 1431010442
-licenseType: Store
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
