--- NewTank/Assets/Tanks/Scripts/Managers.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 6cc27d206e06258438dbc4bb795885b9
-folderAsset: yes
-timeCreated: 1431010007
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
