--- NewTank/Assets/Tanks/Scripts/Managers/GameManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 18e521b6f8588c44f9e68c769421eb47
-timeCreated: 1432810110
-licenseType: Store
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
