--- NewTank/Assets/Tanks/Scripts/Managers/TankManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 1be963428ec33174f88161c5980f3faa
-timeCreated: 1431010478
-licenseType: Store
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
