--- NewTank/Assets/Tanks/Scripts/Shell.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: ef94ef5de0736094c917c0e0217691b6
-folderAsset: yes
-timeCreated: 1431010017
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
