--- NewTank/Assets/Tanks/Scripts/Shell/ShellExplosion.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 4a77fdcebbe85f14ab98dfad5da07c20
-timeCreated: 1432719180
-licenseType: Store
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
