--- NewTank/Assets/Tanks/Scripts/Tank.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 46e032e6247865441876c868e03138e3
-folderAsset: yes
-timeCreated: 1431010026
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
