--- NewTank/Assets/Tanks/Scripts/Tank/TankHealth.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 8326b9dca06e5bc4885942c9371f51f1
-timeCreated: 1431010517
-licenseType: Store
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
