--- NewTank/Assets/Tanks/Scripts/Tank/TankMovement.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 29af3199e11b72e47b41ff380db540fd
-timeCreated: 1431010528
-licenseType: Store
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
