--- NewTank/Assets/Tanks/Scripts/Tank/TankShooting.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 06b38fa3f5188f6419ab17bc1669c18c
-timeCreated: 1431010538
-licenseType: Store
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
