--- NewTank/Assets/Tanks/Scripts/UI.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 22c7d7ea0e44ffc46936169d62a32f92
-folderAsset: yes
-timeCreated: 1431010033
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
