--- NewTank/Assets/Tanks/Scripts/UI/UIDirectionControl.cs
@@ -1,22 +1,0 @@
-using UnityEngine;
-
-public class UIDirectionControl : MonoBehaviour
-{
-    public bool m_UseRelativeRotation = true;  
-
-
-    private Quaternion m_RelativeRotation;     
-
-
-    private void Start()
-    {
-        m_RelativeRotation = transform.parent.localRotation;
-    }
-
-
-    private void Update()
-    {
-        if (m_UseRelativeRotation)
-            transform.rotation = m_RelativeRotation;
-    }
-}
