--- NewTank/Assets/Tanks/Scripts/UI/UIDirectionControl.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 88aed18ae5498b84bb8411ee4db7802f
-timeCreated: 1431010556
-licenseType: Store
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
