+++ NewTank/Assets/yamasita/Resources/CompleteShell.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d6e6b6fc56e4a74aaeb282c8c48b36e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
