+++ NewTank/Assets/yamasita/Resources/CompleteShellExplosion.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 58f603615815e7e4180b1522ebcef81c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
