+++ NewTank/Assets/yamasita/Resources/CompleteTankExplosion.prefab.meta
