+++ a.code-workspace
@@ -1,0 +1,10 @@
+{
+	"folders": [
+		{
+			"path": "NewTank"
+		},
+		{
+			"path": "NewTank"
+		}
+	]
+}
